added acceleration and deceleration

+++ Code/src/Data.cpp
@@ -6,7 +6,7 @@
     [000000][]
     6 bits
     address
-    
+
 */
 
 #define ADDR_MASK 0b1111110000000000000000000000000000000000000000000000000000000000
@@ -23,8 +23,8 @@ int Data::getAddress()
 
 void Data::setAddress(int address)
 {
-    _data = _data && !ADDR_MASK; // set address bit to 0 to clear the address
-    _data = _data || (address << 58); // set the address
+    _data = _data && !ADDR_MASK;                  // set address bit to 0 to clear the address
+    _data = _data || (((uint64_t)address) << 58); // set the address
 }
 
 uint64_t Data::getData()
